requirements.txt
loguru
qrcode[pil]
Pillow
pydantic>=2.0
python-dotenv
aiohttp
httpx